+++ src/avr/src/state.h
@@ -48,6 +48,7 @@ typedef enum {
   HOLD_REASON_PROGRAM_PAUSE,
   HOLD_REASON_OPTIONAL_PAUSE,
   HOLD_REASON_SWITCH_FOUND,
+  HOLD_REASON_SWITCH_NOT_FOUND,
 } hold_reason_t;
 
 
@@ -66,7 +67,7 @@ state_t state_get();
 bool state_is_flushing();
 bool state_is_resuming();
 
-void state_seek_hold();
+void state_seek_hold(bool found);
 void state_holding();
 void state_running();
 void state_jogging();